// src/app/components/search/enums/error.enum.ts
export enum ERROR_MESSAGES {
  COURTROOM = 'The court room number you have entered is not a recognised number for this courthouse. ' +
    'Check and try again. ',
  FUTURE_DATE = 'You have selected a date in the future. The hearing date must be in the past. ',
  START_DATE_MISSING = 'You have not selected a start date.  Select a start date to define your search. ',
  END_DATE_MISSING = 'You have not selected an end date.  Select an end date to define your search. ',
  INVALID_DATE = 'You have not entered a recognised date Enter a date in the correct format. ' +
    'For example, 31/01/2023. ',
}